Teach wallet about cold storage address during withdrawal
I need this change in order to best fix issue #20.

This is basically a modernized revert of commit
aea2a7f6f9b8f5b29424a15bc42d9e9b799a73fb.

During review of #28 we realized this change would write privkeys to
wallet.dat, which might be a security issue, so I reverted it. Later
we realized that the create-deposit-data flow already does that, so I
think this concern is moot.

Also if we should change to a PSBT flow it will require writing the
keys to wallet.dat, so that `walletprocesspsbt` can sign.

Note: This requires Bitcoin Core v0.18.0rc1 or later.