New tests to show issue with validation of extra keys
The Glacier Protocol's deposit protocol recommends a test deposit and
test withdrawal using all N keys. Unfortunately as this failing
wrong-keys test shows, it does not actually validate that each key
belongs to this wallet. If the user enters a key that belongs to
another wallet, it will happily accept it, so long as at least M of
the keys belong to this wallet.

This could give a user false confidence that the key they entered is
good, and potentially lead to loss of funds when the user later tries
to withdraw using that key.

Filed issue: https://github.com/GlacierProtocol/GlacierProtocol/issues/20