Use Decimal everywhere instead of float
So there's no chance of floating point rounding issues.

When bitcoin-cli sends us JSON with floating point amounts, we now
read those as Decimal instead of float. When we dump JSON to
bitcoin-cli, we now dump them as strings (which bitcoin-cli is happy
to accept) just so there's no chance of any floating point precision
issues.

Not certain that I got every instance of floats.