Change bitcoind datadir for developer tests
Previously the datadir was under each testrun/<testname> dir. The
problem was, when I'd change the tests and the RPC port numbers would
change, the Makefile wouldn't be able to kill the old, stale bitcoind
instances because the port number and datadir didn't match
anymore. I'd have to manually `killall bitcoind`.

With this new system, the datadir doesn't depend on the test name but
only on the RPC port number. So even when the port<->test mapping
changes, everything will still work.