New script to run bitcoind in -regtest mode
I will use this to simulate an online system with a blockchain.

Two benefits: (1) I can demonstrate that the withdrawal transactions
created by glacierscript.py are valid, by mining them into the regtest
blockchain; and (2) this will enable me (in the near future) to test
online scripted operations, which will be needed if we switch to using
PSBT with a full node.

I will be converting all the developer tests to use -regtest instead
of -testnet over the next few days.