Extract function, so I can reuse it soon
This introduces a mistake where I call `getnewaddress` twice; I will
fix that in a future commit but it actually serves a useful purpose
(testing the "obsolete" transaction flag in tx.json) later.

Because of that, I don't want to rebase it out of existence. Also, it
would be a major pain.