Use new regtest transaction in this test
The previous transaction was not reproducible; every time I'd rerun
`t/online-regtest-wallet start` I would get a different input
xact. That will be a problem when I try to send the
Glacier-constructed withdrawal transaction to the blockchain (coming
soon). Now that's been fixed, and this input transaction will always
appear in the regtest blockchain.