Change the submit subcommand to do more
Previously this would only verify a provided transaction. The problem
is, where do we find the transaction? We need to scan the golden file
to find it (like the soon-to-be-obsolete decode-transactions script
does).

Since we need to scan for that somewhere, we might as well do that
here, and since we have it here, we might as well make the *.decoded
file here too.

So this will scan the specified golden file, find the Glacier-
generated transaction, submit it to the "online" bitcoind to verify
that it is accepted, then decode it and write it to the *.decoded
file. (At least, when I finish the implementation soon.)

And also soon, the Makefile will run this for every withdrawal.