Use calculated amount for input^2
This changes the input amount used in the p2sh-segwit test to be just
barely higher than the min fee. This is a step towards automating the
creation of all these input transactions in the regtest
blockchain. There isn't enough information in the *.run input
transaction to determine that the input's one input (input^2) had a
value of 1.00000000 btc. It's now slightly less.