Use -txindex for online regtest daemon
This way, when we mine a transaction, it won't hurt to have multiple
parallel processes all mining blocks. Without this change, our call to
`getrawtransaction` would occasionally fail, because while we expected
our subject transaction to be mined in the one block we just mined,
actually it had already been mined in an earlier block by a parallel
submit process.