Reduce fee on test transaction so we don't have to create dust
Recently, Bitcoin Core started enforcing "standard transaction" rules
on regtest. These rules had previously not been enforced on regtest
(and have never been, and still are not, enforced on testnet).

Since that change, attempting to send the transaction built by this
test to the regtest blockchain using `sendrawtransaction` will result
in the transaction being rejected as dust, which causes this test to
fail.

That's now fixed by reducing the fee rate and increasing the output
amount.