Die if not enough keys provided to complete transaction
Rather than give an easily-missed "False" message and still create a
good-looking raw transaction.

The only use case for signing with fewer than M-of-N keys is
sequential signing, and GlacierScript has never supported such a
flow. There's no way, for example, to import a partly-signed raw
transaction and add another signature to it.

It would be fairly easy to support such a flow with PSBT, though.

Also adds new test case to demonstrate this error.