Calculate and display final fee rate after signing
The total fee has already been shown and confirmed, but we should also
show the fee rate.

For create-withdrawal-data, this should exactly match the requested
fee rate as entered by the user, so it's not that interesting.

For PSBT, we can't easily calculate the fee rate before signing,
because we don't know how big the final transaction will be until
after the signatures are added.