WIP: sign maliciously malformed PSBT
An attacker has modified the destination address to send the coins to
himself.

This test fails because `combinepsbt` fails, as expected, but
online_regtest_wallet.py doesn't expect that, and barfs.

I've satisfied myself that `combinepsbt` properly catches this, and
I'm not sure I want to go to all the trouble to handle this in
online_regtest_wallet.py since it's testing Bitcoin Core more than
it's testing Glacier.