Explicitly set replaceable=false in regtest node
This is the default in Bitcoin Core 23, but that changes in 24, and
changes some of our expected transactions. Keep it from changing by
forcing it off all the time.

There's no reason it couldn't be replaceable=true, but we'd have to
redo all the UTXOs in tx.json again if we change it.

Note: this doesn't affect the withdrawal transactions created by
GlacierScript.