Explicitly set replaceable=false for create-withdrawal-data
This is the default in Bitcoin Core 23, but that changes in 24, and
changes some of our withdrawal transactions. Keep it from changing by
forcing it off all the time.

There's no reason it couldn't be replaceable=true, but we'd have to
redo all the expected transactions in the create-withdrawal-data
tests.

Note that for sign-psbt, replaceable is determined by the input PSBT,
and this change does not affect that.